added a MultiSectorTile abstract class valve and hv pump new extend from this class
