+++ src/main/java/org/simulator/board/Components/HVPump.java
@@ -2,7 +2,7 @@ package org.simulator.board.Components;
 
 import org.simulator.board.StateType;
 
-public class HVPump extends ActiveTile{
+public class HVPump extends MultiSectorTile{
     public HVPump(int xCoordinate, int yCoordinate){
         super(StateType.HV_PUMP,"on","off");
         this.xCoordinate = xCoordinate;
