+++ src/main/java/org/simulator/board/Components/PumpingStand.java
@@ -2,8 +2,8 @@ package org.simulator.board.Components;
 
 import org.simulator.board.StateType;
 
-public class PumpStand extends ActiveTile{
-    public PumpStand(int xCoordinate, int yCoordinate){
+public class PumpingStand extends ActiveTile{
+    public PumpingStand(int xCoordinate, int yCoordinate){
         super(StateType.PUMP_STAND,"on","off");
         this.xCoordinate = xCoordinate;
         this.yCoordinate = yCoordinate;
