+++ src/main/java/org/simulator/board/Components/VentingValve.java
@@ -4,7 +4,7 @@ import org.simulator.board.StateType;
 
 public class VentingValve extends ActiveTile{
     public VentingValve(int xCoordinate, int yCoordinate){
-        super(StateType.VENTING_VALVE,"open","close");
+        super(StateType.VENTING_VALVE,"open","closed");
         this.xCoordinate = xCoordinate;
         this.yCoordinate = yCoordinate;
         initializeConnections();
