+++ src/main/java/org/simulator/board/StateType.java
@@ -10,8 +10,8 @@ public enum StateType {
     PIPE(2,4,new Color(127,255,0), false, Pipe.class,"Pipe"),
     CHAMBER(1,4,new Color(0,128,0), false, Chamber.class,"Chamber"),
     GAUGE(1,1,new Color(138,43,226), false, Gauge.class,"Gauge"),
-    HV_PUMP(2,2,Color.YELLOW, true, HVPump.class,"H.V. Pump"),
-    PUMP_STAND(1,1,Color.ORANGE, true, PumpStand.class,"Pump stand"),
+    HV_PUMP(2,2,new Color(255, 255, 0), true, HVPump.class,"H.V. Pump"),
+    PUMP_STAND(1,1,new Color(255, 191, 0), true, PumpingStand.class,"Pumping stand"),
     FORE_VACUUM_PUMP(1,1,new Color(191,105,0), true, ForeVacuumPump.class,"Fore vacuum pump"),
     VALVE(2,2,new Color(70,130,180), true, Valve.class,"Valve"),
     VENTING_VALVE(1,1,new Color(0,0,139), true, VentingValve.class,"Venting valve"),
